pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "authlib>=1.6.0",
    "bcrypt>=4.3.0",
    "boto3>=1.39.3",
    "debugpy>=1.8.14",
    "email-validator>=2.2.0",
    "fastapi>=0.115.14",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.3",
]
